requirements.txt
google-cloud-bigquery[pandas]
google-cloud-monitoring
